requirements.txt
secure-smtplib==0.1.1
requests==2.31.0
plyer==2.1.0
slack-sdk==3.26.1
python-dotenv==1.0.0